# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CyberEd API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/2fa/enroll:
    post:
      operationId: enroll2fa
      tags: [auth]
      summary: Begin 2FA enrollment — returns provisioning URI and QR code
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TotpEnrollResponse"
        "401":
          description: Unauthorized
  /auth/2fa/confirm:
    post:
      operationId: confirm2fa
      tags: [auth]
      summary: Confirm 2FA enrollment using first TOTP code
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TotpCodeInput"
      responses:
        "200":
          description: OK — returns backup codes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TotpConfirmResponse"
        "400":
          description: Bad request
        "401":
          description: Unauthorized
  /auth/2fa/verify:
    post:
      operationId: verify2fa
      tags: [auth]
      summary: Verify TOTP code during login (step 2)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TotpVerifyInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
  /auth/2fa/disable:
    post:
      operationId: disable2fa
      tags: [auth]
      summary: Disable 2FA
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TotpCodeInput"
      responses:
        "204":
          description: No content
        "401":
          description: Unauthorized
...
components:
  schemas:
...
    AuthResponse:
      type: object
      required: [user]
      properties:
        token:
          type: string
          nullable: true
        refreshToken:
          type: string
          nullable: true
        totpRequired:
          type: boolean
        totpChallenge:
          type: string
          description: Temporary token to complete 2FA step
          nullable: true
        user:
          $ref: "#/components/schemas/User"
    TotpEnrollResponse:
      type: object
      required: [provisioningUri, qrCode, secret]
      properties:
        provisioningUri:
          type: string
        qrCode:
          type: string
          description: base64-encoded QR code image
        secret:
          type: string
          description: Manual entry secret
    TotpCodeInput:
      type: object
      required: [code]
      properties:
        code:
          type: string
    TotpConfirmResponse:
      type: object
      required: [backupCodes]
      properties:
        backupCodes:
          type: array
          items:
            type: string
    TotpVerifyInput:
      type: object
      required: [totpChallenge, code]
      properties:
        totpChallenge:
          type: string
        code:
          type: string
...
    User:
      type: object
      required: [id, username, email, role, createdAt]
      properties:
        id:
          type: number
        username:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [admin, editor, viewer]
        createdAt:
          type: string
          format: date-time
        emailVerifiedAt:
          type: string
          format: date-time
          nullable: true
...